+++ src/players/fibonacci.py
